fix(ux): cast average score to float for Unity compile

+++ Assets/Scripts/BoardHandler.cs
@@ -373,7 +373,7 @@ public class BoardHandler : MonoBehaviour
         if (points.Count == 0)
             return 0f;
 
-        return points.Average();
+        return (float)points.Average();
     }
 
     public float GetAverageStability()
